==> 04.Metabolomics.Rmd <==
---
title: "04.Metabolomics"
author: "Kelsey Huus"
date: "2024-07-29"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
##Packages
library(dplyr)
library(ggplot2)
#library(data.table)
library(tidyr)
library(Maaslin2)
library(ggrepel)
library(pheatmap)
library(stringr)

```


```{r import1}
#import metadata
#updated: this version also excludes SERIES 2 of all double series.
meta_seq <- read.csv("/ebio/abt3_projects2/uHEAT/data/metadata/meta_seq_clean_series1.24.08.18.csv", 
                     row.names=1)
meta_seq <- meta_seq %>% filter(!is.na(fever7d_aboveavg))
length(meta_seq$SampleID)

#define DIET 3
meta_seq$Diet3 <- ifelse(meta_seq$Diet %in% c("Vegan", "Vegetarian"), yes="Veg", no="Omni")
meta_seq$Diet3 <- factor(meta_seq$Diet3, levels=c("Veg", "Omni"))

#per-person metadata, matching
meta <- meta_seq %>% filter(!duplicated(Participant_ID)&!is.na(fever7d_aboveavg)) %>%
  dplyr::select(-c(SampleID, Relative_Day, Time0, Time01))
length(meta$Participant_ID) #final number of participants

meta$fever7d_aboveavg <- factor(meta$fever7d_aboveavg, levels=c("lo", "hi"))

```


```{r triglycerides}
###Total TG inferred from Dai Long
tg <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/MSF-SmallMoleculeTransitionResults_TG_converted.csv")
names(tg) <- gsub("_Pos", "",names(tg), fixed=TRUE)
names(tg)

mdp <- data.frame(Metabo_SampleID=c(names(tg)))
mdp$Participant_ID=sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 2)
mdp$Participant_ID <- gsub("H", "HEAT_", mdp$Participant_ID)
mdp$Blood <- sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 3)
mdp$Blood <- ifelse(mdp$Participant_ID %in% c("Pos", "Neg"), yes=NA, no=mdp$Blood)
mdp$Participant_ID <- ifelse(mdp$Participant_ID%in%c("Pos", "Neg"), yes=NA, no=mdp$Participant_ID)
mdp$SampleType <- sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 1)
mdp$SampleType <- ifelse(mdp$SampleType %in% c("Blank", "QC"), yes=mdp$SampleType, no="Serum")
head(mdp)
meta <- meta_seq %>% filter(!duplicated(Participant_ID))
mdp <- mdp %>% full_join(meta) %>% filter(!is.na(SampleType)) #keep full metadata for all the samples
head(mdp)
row.names(mdp) <- mdp$Metabo_SampleID

#and now I have a way to refer between metadata and metabo ID, at least.
head(tg)
#I really wanted a summary across all lipids in those categories
tg$Replicate.Name
tg$Category <- sapply(strsplit(as.character(tg$Replicate.Name)," "), `[`, 1)
#summarize for all the people though? make it long first I guess.
tg_long <- pivot_longer(tg, cols=c(2:249), names_to='Metabo_SampleID', values_to='Abundance')
tg_long
tg_sum <- tg_long %>% group_by(Metabo_SampleID, Category) %>% 
  dplyr::summarize(Abundance=sum(Abundance))
tg_sum
#pivot back wider
tg_wide <- tg_sum %>% pivot_wider(names_from=Category, values_from=Abundance)
tg_wide
#OK, and then combine with the metadata.
tg_sum_meta <- full_join(tg_wide, mdp) %>% filter(Blood=="B1"&!is.na(fever7d_aboveavg)) #baseline only #only clean samples

#write out the first columns to combine with things later
#write.csv(tg_sum_meta[,c(2:6)], "/ebio/abt3_projects2/uHEAT2/data/metabolomics/triglycerides_baseline_cleaned.csv")

#and thennn plot.
ggplot(tg_sum_meta, aes(x=Diet2, y=CE)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$CE~tg_sum_meta$Diet2) #p=0.1
ggplot(tg_sum_meta, aes(x=Diet2, y=TG)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$TG~tg_sum_meta$Diet2) #p=0.02 - higher in Omni as expected
ggplot(tg_sum_meta, aes(x=Diet2, y=`1,2-DG`)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$`1,2-DG`~tg_sum_meta$Diet2) #p=0.0008 - higher in Omni as expected
ggplot(tg_sum_meta, aes(x=Diet2, y=`1,3-DG`)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$`1,3-DG`~tg_sum_meta$Diet2) #p=0.0005 - higher in Omni as expected

#vs Diet 3
wilcox.test(tg_sum_meta$TG~tg_sum_meta$Diet3) #p=0.02 - higher in Omni as expected
wilcox.test(tg_sum_meta$CE~tg_sum_meta$Diet3) #p=0.1
wilcox.test(tg_sum_meta$`1,2-DG`~tg_sum_meta$Diet3) #p=0.0004 - higher in Omni as expected
wilcox.test(tg_sum_meta$`1,3-DG`~tg_sum_meta$Diet3) #p=0.0003 - higher in Omni as expected

#cool and what about by fever, then?
ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=CE)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$CE~tg_sum_meta$fever7d_aboveavg) #p=0.6
ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=TG)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$TG~tg_sum_meta$fever7d_aboveavg) #p=0.12 
ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=`1,2-DG`)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$`1,2-DG`~tg_sum_meta$fever7d_aboveavg) #p=0.1
ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=`1,3-DG`)) + geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0)
wilcox.test(tg_sum_meta$`1,3-DG`~tg_sum_meta$fever7d_aboveavg) #p=0.07

m1 <- lm(`1,3-DG`~Sex + Age + Diet2 + fever7d_aboveavg, data=tg_sum_meta)
summary(m1)

#total tri- and di-glycerides?
tg_sum_meta$total_G <- tg_sum_meta$TG + tg_sum_meta$`1,2-DG` + tg_sum_meta$`1,3-DG`
wilcox.test(tg_sum_meta$total_G~tg_sum_meta$fever7d_aboveavg) #p=0.1
wilcox.test(tg_sum_meta$total_G~tg_sum_meta$Diet3) #p=0.02

#are there any correlations with BMI
cor.test(tg_sum_meta$BMI_kg_m2, tg_sum_meta$CE, method='spearman')
ggplot(tg_sum_meta, aes(x=BMI_kg_m2, y=CE)) + geom_point() + geom_smooth(method='lm')
cor.test(tg_sum_meta$BMI_kg_m2, tg_sum_meta$TG, method='spearman') #yes very significant
ggplot(tg_sum_meta, aes(x=BMI_kg_m2, y=TG)) + geom_point() + geom_smooth(method='lm')
cor.test(tg_sum_meta$BMI_kg_m2, tg_sum_meta$`1,2-DG`, method='spearman') #yes very significant
ggplot(tg_sum_meta, aes(x=BMI_kg_m2, y=`1,2-DG`)) + geom_point() + geom_smooth(method='lm')
cor.test(tg_sum_meta$BMI_kg_m2, tg_sum_meta$`1,3-DG`, method='spearman') #yes very significant
ggplot(tg_sum_meta, aes(x=BMI_kg_m2, y=`1,3-DG`)) + geom_point() + geom_smooth(method='lm')

#plot these out nicely by both diet and fever
tg_sum_meta$fever7d_aboveavg <- factor(tg_sum_meta$fever7d_aboveavg, levels=c("lo", "hi"))
tg_sum_meta$Diet3 <- factor(tg_sum_meta$Diet3, levels=c("Veg", "Omni"))

#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/DietFever_TG.pdf")
p <- ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=TG)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  facet_grid(.~Diet3) +
  ylab("Triglycerides") + xlab("Fever") +
  annotate("segment", x=1, xend=2, y=2.5e+09, yend=2.5e+09) + 
  annotate("text", x=1.5, y=2.6e+09, label=c("ns"))
p
#dev.off()

wilcox.test(tg_sum_meta$TG~tg_sum_meta$Diet3)
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/S3C_DietTG.pdf")
p <- ggplot(tg_sum_meta, aes(x=Diet3, y=TG)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  ylab("Triglycerides") + xlab("Diet") +
  annotate("segment", x=1, xend=2, y=2.55e+09, yend=2.55e+09) +
  annotate("text", x=1.5, y=2.6e+09, label=c("*"), size=6)
p
#dev.off()

hist(tg_sum_meta$TG) #quite normal already
#two-way anova
#a2 <- aov(TG~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=tg_sum_meta) #sig or near sig, for both factors, but not interaction
#summary(a2)

#pairwise.t.test(tg_sum_meta$TG, g=paste(tg_sum_meta$fever7d_aboveavg, tg_sum_meta$Diet3, sep="_"),
#                p.adjust.method = "BH")
tg_sum_meta$DietFever <- paste(tg_sum_meta$Diet3, tg_sum_meta$fever7d_aboveavg, sep="_")
a2 <- aov(TG~DietFever, data=tg_sum_meta)
posthoc <- TukeyHSD(a2)


#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/plots5/DietFever_12DG.pdf")
p <- ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=`1,2-DG`)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  facet_grid(.~Diet3) +
  ggtitle("1,2-Diglycerides") + xlab("Fever")
p
#dev.off()

wilcox.test(tg_sum_meta$`1,2-DG`~tg_sum_meta$Diet3)
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/S3D_Diet12DG.pdf")
p <- ggplot(tg_sum_meta, aes(x=Diet3, y=`1,2-DG`)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  ylab("1,2-Diglycerides") + xlab("Diet") +
  annotate("segment", x=1, xend=2, y=4.4e+07, yend=4.4e+07) +
  annotate("text", x=1.5, y=4.5e+07, label=c("***"), size=6)
p
#dev.off()

hist(tg_sum_meta$`1,2-DG`) #quite normal already
#two-way anova
a2 <- aov(`1,2-DG`~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=tg_sum_meta) #only diet
summary(a2)

#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/plots5/DietFever_13DG.pdf")
p <- ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=`1,3-DG`)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  facet_grid(.~Diet3) +
  ggtitle("1,3-Diglycerides") + xlab("Fever")
p
#dev.off()

wilcox.test(tg_sum_meta$`1,3-DG`~tg_sum_meta$Diet3)
pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/S3E_Diet13DG.pdf")
p <- ggplot(tg_sum_meta, aes(x=Diet3, y=`1,3-DG`)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  ylab("1,3-Diglycerides") + xlab("Diet") +
  annotate("segment", x=1, xend=2, y=1.8e+07, yend=1.8e+07) +
  annotate("text", x=1.5, y=1.9e+07, label=c("***"), size=6)
p
dev.off()

#CE
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/plots5/DietFever_CE.pdf")
p <- ggplot(tg_sum_meta, aes(x=fever7d_aboveavg, y=`CE`)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  facet_grid(.~Diet3) +
  ggtitle("Cholesterol Esters") + xlab("Fever")
p
#dev.off()

hist(tg_sum_meta$`1,3-DG`) #quite normal already
#two-way anova
a2 <- aov(`1,3-DG`~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=tg_sum_meta) #diet and also marginally fever
summary(a2)

#linear models
m1 <- lm(`1,3-DG`~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=tg_sum_meta)
summary(m1) #only diet

m2 <- lm(`1,2-DG`~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=tg_sum_meta)
summary(m2) #only diet

m2 <- lm(TG~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=tg_sum_meta)
summary(m2) #only diet

```

```{r cholesterol}
cholesterol <- read.csv("/ebio/abt3_projects2/uHEAT2/data/cholesterol/Cholesterol_interpolations_all.csv")
cholesterol$Participant_ID <- sapply(strsplit(cholesterol$Sample,"_"), `[`, 1)
cholesterol$Participant_ID <- gsub("H", "HEAT_", cholesterol$Participant_ID)
cholesterol$Cholesterol_Type <- sapply(strsplit(cholesterol$Sample,"_"), `[`, 3)
head(cholesterol)

#rearrange
cholesterol2 <- cholesterol %>% 
  dplyr::select(c(Participant_ID, Cholesterol_Type, Plate, Cholesterol_mg.dL)) %>%
  filter(Participant_ID!="blank") %>%
  pivot_wider(names_from=Cholesterol_Type, values_from=Cholesterol_mg.dL) %>%
  rename(HDL_Cholesterol_mg.dL=HDL, LDL_Cholesterol_mg.dL=LDL) %>%
  mutate(Total_Cholesterol_mg.dL=HDL_Cholesterol_mg.dL+LDL_Cholesterol_mg.dL)
head(cholesterol2)
tail(cholesterol2)
#there are a few technical duplicates, take the first (non duplicated) value
cholesterol2 <- cholesterol2 %>% filter(!duplicated(Participant_ID))

cholesterol2$Cholesterol_ratio <- cholesterol2$LDL_Cholesterol_mg.dL / cholesterol2$HDL_Cholesterol_mg.dL
#export this version to combine with Table S1, etc.
#write.csv(cholesterol2, "/ebio/abt3_projects2/uHEAT2/data/cholesterol/Cholesterol_clean.csv" )

#add metadata
cholesterol_meta <- full_join(cholesterol2, meta) %>% 
  filter(!is.na(Total_Cholesterol_mg.dL)&!duplicated(Participant_ID)&!is.na(fever7d_aboveavg))

#test by fever
pdata2 <- cholesterol_meta %>% filter(!is.na(fever7d_aboveavg))
pdata2$fever7d_aboveavg <- factor(pdata2$fever7d_aboveavg, levels=c("lo", "hi"))

wilcox.test(pdata2$Total_Cholesterol_mg.dL~pdata2$fever7d_aboveavg) #p=0.86
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT2/data/cholesterol/Fever_totalCholesterol_plates1-10.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=Total_Cholesterol_mg.dL)) + geom_boxplot() + geom_jitter(width=0.1, height=0) 
p <- p + theme_bw()
p <- p + annotate('segment', x=1, xend=2, y=2.5, yend=2.5)
p <- p + annotate('text', label=c("NS"), x=1.5, y=2.6)
p
#dev.off()

wilcox.test(pdata2$LDL_Cholesterol_mg.dL~pdata2$fever7d_aboveavg) #p=0.36
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT2/data/cholesterol/Fever_LDLCholesterol_plates1-10.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=LDL_Cholesterol_mg.dL)) + geom_boxplot() + geom_jitter(width=0.1, height=0)
p <- p + theme_bw()
p <- p + annotate('segment', x=1, xend=2, y=2, yend=2)
p <- p + annotate('text', label=c("NS"), x=1.5, y=2.1)
p
#dev.off()

wilcox.test(pdata2$HDL_Cholesterol_mg.dL~pdata2$fever7d_aboveavg) #p=0.136 #hmph.
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT2/data/cholesterol/Fever_HDLCholesterol_plates1-10.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=HDL_Cholesterol_mg.dL)) + geom_boxplot() + geom_jitter(width=0.1, height=0)
p <- p + theme_bw()
p <- p + annotate('segment', x=1, xend=2, y=1, yend=1)
p <- p + annotate('text', label=c("p=0.13"), x=1.5, y=1.1)
p
#dev.off()

wilcox.test(pdata2$Cholesterol_ratio~pdata2$fever7d_aboveavg) #p=0.02 #indeed
#cairo_pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/2C.CholesterolRatio.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=Cholesterol_ratio, colour=fever7d_aboveavg)) + 
  geom_boxplot(outlier.shape=NA) + geom_jitter(width=0.1, height=0) + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) + 
  annotate('segment', x=1, xend=2, y=6, yend=6) + 
  annotate('text', label=c("*"), x=1.5, y=6.2) + 
  scale_colour_manual(values=c("grey50", "firebrick3")) + 
  ylab("Ratio LDL/HDL") + xlab("Fever") + 
  theme(legend.position="none")
p
#dev.off()

#test by diet for comparison
wilcox.test(pdata2$Total_Cholesterol_mg.dL~pdata2$Diet3) #NS
wilcox.test(pdata2$HDL_Cholesterol_mg.dL~pdata2$Diet3) #p=0.1
wilcox.test(pdata2$LDL_Cholesterol_mg.dL~pdata2$Diet3) #p=0.2
wilcox.test(pdata2$Cholesterol_ratio~pdata2$Diet3) #p=0.002

#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/S3F_Diet3_CholesterolRatio.pdf")
p <- ggplot(pdata2, aes(x=Diet3, y=Cholesterol_ratio)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  annotate('segment', x=1, xend=2, y=6, yend=6) + 
  annotate('text', label=c("**"), x=1.5, y=6.2, size=6) + 
  ylab("Cholesterol LDL:HDL Ratio")  + xlab("Diet")
p
#dev.off()

#and is it entirely driven by diet? #no, the ratio is still worse in fever-hi within each category

#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/DietFever_CholesterolRatio.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=Cholesterol_ratio)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  annotate('segment', x=1, xend=2, y=6, yend=6) + 
  annotate('text', label=c("ns"), x=1.5, y=6.2) + 
  ylab("Cholesterol LDL:HDL Ratio") + xlab("Fever") +
  facet_grid(.~Diet3)
p
#dev.off()

hist(log(pdata2$Cholesterol_ratio)) #quite normal when log adjusted
#two-way anova
a2 <- aov(log(Cholesterol_ratio)~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=pdata2) #sig, for both factors, but not interaction
summary(a2)
t <- TukeyHSD(a2)
t #but it only reaches significance for low-veg vs hi-omni, like TG.

#this is basically the same thing, btw.
pdata2$DietFever <- paste(pdata2$Diet3, pdata2$fever7d_aboveavg, sep="_")
a2b <- aov(log(Cholesterol_ratio)~DietFever, data=pdata2)
posthoc <- TukeyHSD(a2b)


m2 <- lm(Cholesterol_ratio~fever7d_aboveavg + Diet3 + fever7d_aboveavg:Diet3, data=pdata2)
summary(m2) #only diet

#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/plots5/DietFever_TotalCholesterol.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=Total_Cholesterol_mg.dL)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  #annotate('segment', x=1, xend=2, y=6, yend=6) + 
  #annotate('text', label=c("**"), x=1.5, y=6.2, size=5) + 
  ylab("Ratio LDL/HDL") + ggtitle("Cholesterol (mg/dL)") + xlab("Fever") +
  facet_grid(.~Diet3)
p
#dev.off()

#test by IgG hi/lo incase it's stronger in IgG lo
pdata2$IgG_B1_aboveavg <- factor(pdata2$IgG_B1_aboveavg, levels=c("lo", "hi"))
#pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/0G_FeverChol_ByIgGLo.pdf")
p <- ggplot(pdata2, aes(x=fever7d_aboveavg, y=Cholesterol_ratio, colour=fever7d_aboveavg)) + geom_boxplot(outlier.shape=NA) + 
  geom_jitter(width=0.1, height=0) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) +
  ylab("Cholesterol LDL:HDL Ratio") + xlab("Fever") + theme(legend.position='none') +
    scale_colour_manual(values=c("grey60", "firebrick3")) +
  facet_grid(.~IgG_B1_aboveavg)
p
#dev.off()


```

```{r calpro}
lc <- read.csv("/ebio/abt3_projects2/uHEAT/data/metadata/LCN2.Calpro.ELISA.Interpolations.csv")
head(lc)
#multiple both LCN2 and Calpro by 20 for dilution factor 
#note that where the factor is different (plate 5) due to a 1:10 instead of 1:20 dilution, this is already manually adjusted by 1/2 in the excel sheet
lc$LCN2 <- lc$LCN2*20
lc$Calpro <- lc$Calpro*20
#for the few duplicates, take an average per participant 
lc$Participant_ID <- sapply(strsplit(lc$SampleID,"_"), `[`, 1)
lc$Participant_ID <- gsub("H", "HEAT_", lc$Participant_ID)
lc <- lc %>% group_by(Participant_ID) %>%
  dplyr::summarize(LCN2.pg.mL=mean(LCN2, na.rm=TRUE),
            Calpro.pg.mL=mean(Calpro, na.rm=TRUE))
lc <- lc %>% filter(Participant_ID!="BLANK")
lc

```

```{r import2 untargeted}

#import metabolomics
pos <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/MS019_pos_final_Nov2024.csv") #updated with higher confidence metabolites
neg <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/MS019_neg_final_Nov2024.csv") #updated with higher confidence metabolites

#per-person metadata, matching
meta <- meta_seq %>% filter(!duplicated(Participant_ID)&!is.na(fever7d_aboveavg)) %>%
  dplyr::select(-c(SampleID, Relative_Day, Time0, Time01))
dim(meta)

#retrieve annotations directly from the data file
names(pos)
dim(pos)
annotations_pos <- pos[,c(1:32)]
names(neg)
dim(neg)
annotations_neg <- neg[,c(1:32)]

#extract the data itself, separately., as I had it before
pos <- pos[,c(1:3, 33:length(names(pos)))]
neg <- neg[,c(1:3, 33:length(names(neg)))]

#make a unified annotation column - pick the first name if it is exists, otherwise the second
annotations_pos[annotations_pos==""] <- NA
annotations_pos$annotation <- ifelse(!is.na(annotations_pos$spectral_db_matches.compound_name),
                                     yes=annotations_pos$spectral_db_matches.compound_name,
                                     no=annotations_pos$name)
#and then if both of those names are missing, use the class.
annotations_pos$annotation <- ifelse(is.na(annotations_pos$annotation),
                                     yes=annotations_pos$ClassyFire.most.specific.class,
                                     no=annotations_pos$annotation)
annotations_pos$annotation
annotations_pos$id <- paste("Mp", annotations_pos$id, sep="_")
#make a small version for merging
ap <- annotations_pos %>% dplyr::select(id, rt, mz, annotation, molecularFormula, 
                                 NPC.pathway, NPC.pathway.Probability, 
                                 NPC.superclass, NPC.superclass.Probability)
ap #and now it is clear that some of them are dups, so only do the annotation merging aftper Maaslin2.
#make a non-numeric id for rownames and column names.


#make a unified annotation column - pick the first name if it is exists, otherwise the second
annotations_neg[annotations_neg==""] <- NA
annotations_neg$annotation <- ifelse(!is.na(annotations_neg$spectral_db_matches.compound_name),
                                     yes=annotations_neg$spectral_db_matches.compound_name,
                                     no=annotations_neg$name)
#and then if both of those names are missing, use the class.
annotations_neg$annotation <- ifelse(is.na(annotations_neg$annotation),
                                     yes=annotations_neg$ClassyFire.most.specific.class,
                                     no=annotations_neg$annotation)
annotations_neg$annotation
annotations_neg$id <- paste("Mn", annotations_neg$id, sep="_")
#make a small version for merging
an <- annotations_neg %>% dplyr::select(id, rt, mz, annotation, molecularFormula, 
                                 NPC.pathway, NPC.pathway.Probability, 
                                 NPC.superclass, NPC.superclass.Probability)
an #and now it is clear that some of them are dups, so only do the annotation merging after Maaslin2.
#make a non-numeric id for rownames and column names.


#cleaning it up cont. #weirdly in the new positive version there are 'duplicate ids'.
row.names(pos) <- paste("Mp", pos$id, sep="_")
ftp <- pos %>% dplyr::select(-c("id", "rt", "mz")) #start with positive ion mode - name based on Petras lab script
ftp <- ftp[,order(colnames(ftp)), drop=F]

row.names(neg) <- paste("Mn", neg$id, sep="_")
ftn <- neg %>% dplyr::select(-c("id", "rt", "mz")) #start with positive ion mode - name based on Petras lab script
ftn <- ftn[,order(colnames(ftn)), drop=F]

#make a 'metabolomics' metadata file which contains all the sample info also for blanks and QCs
#can I do this identically for the neg and pos ones? #technically not but the only diff is pos/neg.
identical(names(ftp), names(ftn)) #OK still not identical even with pos/neg designation removed. just make two separate metadata files for now.
mdp <- data.frame(Metabo_SampleID=c(names(ftp)))
mdp$Participant_ID=sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 2)
mdp$Participant_ID <- gsub("H", "HEAT_", mdp$Participant_ID)
mdp$Blood <- sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 3)
mdp$Blood <- ifelse(mdp$Participant_ID %in% c("Pos", "Neg"), yes=NA, no=mdp$Blood)
mdp$Participant_ID <- ifelse(mdp$Participant_ID%in%c("Pos", "Neg"), yes=NA, no=mdp$Participant_ID)
mdp$SampleType <- sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 1)
mdp$SampleType <- ifelse(mdp$SampleType %in% c("Blank", "QC", "Average", "NA_num"), yes=mdp$SampleType, no="Serum")
#meta_small <- meta %>% dplyr::select(Participant_ID, fever7d_aboveavg) #start with just minimal metadata to keep things simple
mdp <- mdp %>% full_join(meta) %>% filter(!is.na(SampleType)) #keep full metadata for all the samples
head(mdp)
row.names(mdp) <- mdp$Metabo_SampleID
mdp <- mdp[order(row.names(mdp)),,drop=F]
#did any samples still get run twice? I think it is fixed now.
mdp$Metabo_SampleNo=sapply(strsplit(as.character(mdp$Metabo_SampleID),"_"), `[`, 1)
mdp[which(duplicated(mdp$Metabo_SampleNo)),]$Metabo_SampleNo #yes, #71 and #97 are duplicated?
mdp[which(duplicated(mdp$Metabo_SampleNo)),]$Metabo_SampleID
dups <- mdp %>% filter(Metabo_SampleNo %in% c("71", "97"))
ftp[,which(names(ftp) %in% dups$Metabo_SampleID)]
#remove the duplicates from the ftp table according to Dennis' instructions on which one is a better intensity.
mdp <- mdp %>% filter(!(Metabo_SampleID %in% c("71_H065_B2_Pos", "97_H107_B1_Pos")))
mdp[which(duplicated(mdp$Metabo_SampleNo)),]$Metabo_SampleNo #Now it's clean.
ftp <- ftp[,which(names(ftp)%in%mdp$Metabo_SampleID)]

# how many files in the metadata are also present in the feature table
table(rownames(mdp) %in% colnames(ftp))
identical(rownames(mdp), colnames(ftp))
#transpose the feature table and ensure numeric
ftp_t <- as.data.frame(t(ftp)) #transposing the ftp
ftp_t <- ftp_t %>% mutate_all(as.numeric)  #converting all values to numeric
identical(rownames(mdp),rownames(ftp_t)) #should return TRUE now

##repeat for negative mode.
mdn <- data.frame(Metabo_SampleID=c(names(ftn)))
mdn$Participant_ID=sapply(strsplit(as.character(mdn$Metabo_SampleID),"_"), `[`, 2)
mdn$Participant_ID <- gsub("H", "HEAT_", mdn$Participant_ID)
mdn$Blood <- sapply(strsplit(as.character(mdn$Metabo_SampleID),"_"), `[`, 3)
mdn$Blood <- ifelse(mdn$Participant_ID %in% c("Pos", "Neg"), yes=NA, no=mdn$Blood)
mdn$Participant_ID <- ifelse(mdn$Participant_ID%in%c("Pos", "Neg"), yes=NA, no=mdn$Participant_ID)
mdn$SampleType <- sapply(strsplit(as.character(mdn$Metabo_SampleID),"_"), `[`, 1)
mdn$SampleType <- ifelse(mdn$SampleType %in% c("Blank", "QC", "Average", "NA_num"), yes=mdn$SampleType, no="Serum")
#meta_small <- meta %>% dplyr::select(Participant_ID, fever7d_aboveavg) #start with just minimal metadata to keep things simple
mdn <- mdn %>% full_join(meta) %>% filter(!is.na(SampleType)) #keep full metadata for all the samples
head(mdn)
row.names(mdn) <- mdn$Metabo_SampleID
mdn <- mdn[order(row.names(mdn)),,drop=F]
#did any samples still get run twice? I think it is fixed now.
mdn$Metabo_SampleNo=sapply(strsplit(as.character(mdn$Metabo_SampleID),"_"), `[`, 1)
mdn[which(duplicated(mdn$Metabo_SampleNo)),]$Metabo_SampleNo #yes, #149 and 93 and 95 dups #no only blanks and QCs
mdn[which(duplicated(mdn$Metabo_SampleNo)),]$Metabo_SampleID #yes, #149 and 93 and 95 dups
dups <- mdn %>% filter(Metabo_SampleNo %in% c("93", "95", "149"))
ftn[,which(names(ftn) %in% dups$Metabo_SampleID)]
#remove the duplicates from the ftp table according to Dennis' instructions on which one is a better intensity.
#later noticed 2 additional technical duplicates, filtered based on which one had fewer NA
length(ftn$X114_H051_B1_Neg[is.na(ftn$X114_H051_B1_Neg)])
length(ftn$X167_H051_B1_Neg_1[is.na(ftn$X167_H051_B1_Neg_1)]) #has less NA
length(ftn$X154_H104_B1_Neg_1[is.na(ftn$X154_H104_B1_Neg_1)])
length(ftn$X86_H104_B1_Neg_2[is.na(ftn$X86_H104_B1_Neg_2)]) #very similar, has slightly less NA
mdn <- mdn %>% filter(!(Metabo_SampleID %in% c("149_H038_B2_Neg_1", "93_H154_B1_Neg_1", "93_H154_B1_Neg_2", "95_H155_B1_Neg_1",
                                               "X114_H051_B1_Neg", "X154_H104_B1_Neg_1")))
mdn[which(duplicated(mdn$Metabo_SampleNo)),]$Metabo_SampleNo #now it's clean.
ftn <- ftn[,which(names(ftn)%in%mdn$Metabo_SampleID)]

# how many files in the metadata are also present in the feature table
table(rownames(mdn) %in% colnames(ftn))
identical(rownames(mdn), colnames(ftn))
#transpose the feature table and ensure numeric
ftn_t <- as.data.frame(t(ftn)) #transposing the ftn
ftn_t <- ftn_t %>% mutate_all(as.numeric)  #converting all values to numeric
identical(rownames(mdn),rownames(ftn_t)) #should return TRUE now

```

```{r zeroes}
#first, I think that all NAs should be considered zeroes for the purposes of means, prevalence, etc.
ftp_t[is.na(ftp_t)] <- 0
ftn_t[is.na(ftn_t)] <- 0

```

```{r filter}
##BECAUSE DAI LONG HAS NOW MANUALLY FILTERED FOR HIGH-CONFIDENCE / ABUNDANCE METABOLITES: ALSO SKIP.
#POS.
#separate samples and blanks.
blank_ids_p <- mdp[which(mdp$SampleType=="Blank"),]$Metabo_SampleID
blank_ids_p
qc_ids_p <- mdp[which(mdp$SampleType=="QC"),]$Metabo_SampleID
qc_ids_p
sample_ids_p <- mdp[which(mdp$SampleType=="Serum"),]$Metabo_SampleID
sample_ids_p

# Getting the corresponding rows from ftp_t
Blank_p <- ftp_t[which(rownames(ftp_t) %in% (blank_ids_p)), , drop=F]
Samples_p <- ftp_t[which(rownames(ftp_t) %in% (sample_ids_p)), , drop=F]

#Define a cutoff for blank features removal.
#When cutoff is low, more noise (or background) detected; With higher cutoff, less background detected, thus more features observed
Cutoff <- 0.3 # (i.e. 10% - 100%). Ideal cutoff range: 0.1-0.3

#Perform Blank Removal
#Getting mean for every feature in blank and Samples in a data frame named 'Avg_ftp'
#note - this doesn't help for prevalence. Later, we shoudl also filter for prevalence.
Avg_ftp <- data.frame(Avg_blank=colMeans(Blank_p, na.rm = T)) # set na.rm = F to check if there are NA values. When set as T, NA values are changed to 0
Avg_ftp$`Avg_samples` <- colMeans(Samples_p, na.rm= T) # adding another column 'Avg_samples' for feature means of samples
#this still contains many NaN from when the ENTIRE COLUMN is blank.
#all NaNs now replaced by zeroes to begin with

#Getting the ratio of blank vs Sample
Avg_ftp$`Ratio_blank_Sample` <- (Avg_ftp$`Avg_blank`+1)/(Avg_ftp$`Avg_samples`+1)

# Creating a bin with 1s when the ratio>Cutoff (EXCLUDE), else put 0s (INCLUDE)
Avg_ftp$`Bg_bin` <- ifelse(Avg_ftp$`Ratio_blank_Sample` > Cutoff, 1, 0 )
Avg_ftp #there are a lot of NA!!! from things that appear in samples but not at all in blanks.
Blank_p$Mp_3 #for example, this whole thing is entirely NA. Or zero.
#Avg_ftp$Bg_bin <- ifelse(is.na(Avg_ftp$Avg_blank), yes=0, no=Avg_ftp$Bg_bin) #correct the NA blanks to 'do not remove' zeros


#Calculating the number of background features and features present
print(paste("Total no.of features:",nrow(Avg_ftp)))
print(paste("No.of Background or noise features:",sum(Avg_ftp$`Bg_bin` ==1,na.rm = T))) 
print(paste("No.of features after excluding noise:",(ncol(Samples_p) - sum(Avg_ftp$`Bg_bin` ==1,na.rm = T))))

#do this in two steps to avoid lo
identical(names(Samples_p), row.names(Avg_ftp)) #the metabolite ids are the same
blk_rem_p <- merge(as.data.frame(t(Samples_p)), Avg_ftp, by=0) %>%
  filter(Bg_bin == 0) %>% #picking only the features
  dplyr::select(-c(Avg_blank,Avg_samples,Ratio_blank_Sample,Bg_bin)) #removing the last 4 columns
row.names(blk_rem_p) <- blk_rem_p$Row.names #extra step bc columns-to-rownames didn't work
blk_rem_p <- as.data.frame(t(blk_rem_p %>% dplyr::select(-c(Row.names))))
dim(blk_rem_p) #now the samples are retained
names(blk_rem_p) #and so are the metabolite names :)

#NEG
#repeat for neg
#separate samples and blanks.
blank_ids_n <- mdn[which(mdn$SampleType=="Blank"),]$Metabo_SampleID
blank_ids_n
qc_ids_n <- mdn[which(mdn$SampleType=="QC"),]$Metabo_SampleID
qc_ids_n
sample_ids_n <- mdn[which(mdn$SampleType=="Serum"),]$Metabo_SampleID
sample_ids_n

# Getting the corresponding rows from ftp_t
Blank_n <- ftn_t[which(rownames(ftn_t) %in% (blank_ids_n)), , drop=F]
Samples_n <- ftn_t[which(rownames(ftn_t) %in% (sample_ids_n)), , drop=F]

#Define a cutoff for blank features removal.
#When cutoff is low, more noise (or background) detected; With higher cutoff, less background detected, thus more features observed
Cutoff <- 0.3 # (i.e. 10% - 100%). Ideal cutoff range: 0.1-0.3

#Perform Blank Removal
#Getting mean for every feature in blank and Samples in a data frame named 'Avg_ftn'
Avg_ftn <- data.frame(Avg_blank=colMeans(Blank_n, na.rm= T)) # set na.rm = F to check if there are NA values. When set as T, NA values are changed to 0
Avg_ftn$`Avg_samples` <- colMeans(Samples_n, na.rm= T) # adding another column 'Avg_samples' for feature means of samples


#Getting the ratio of blank vs Sample
Avg_ftn$`Ratio_blank_Sample` <- (Avg_ftn$`Avg_blank`+1)/(Avg_ftn$`Avg_samples`+1)

# Creating a bin with 1s when the ratio>Cutoff, else put 0s
Avg_ftn$`Bg_bin` <- ifelse(Avg_ftn$`Ratio_blank_Sample` > Cutoff, 1, 0 )
Avg_ftn #there are a lot of NA!!! from things that appear in samples but not at all in blanks.
#Blank$V1 #for example, this whole thing is entirely NA.
#Avg_ftn$Bg_bin <- ifelse(is.na(Avg_ftn$Avg_blank), yes=0, no=Avg_ftn$Bg_bin) #correct the NA blanks to 'do not remove' zeros
#Avg_ftn

#Calculating the number of background features and features present
print(paste("Total no.of features:",nrow(Avg_ftn)))
print(paste("No.of Background or noise features:",sum(Avg_ftn$`Bg_bin` ==1,na.rm = T))) 
print(paste("No.of features after excluding noise:",(ncol(Samples_n) - sum(Avg_ftn$`Bg_bin` ==1,na.rm = T))))

blk_rem_n <- merge(as.data.frame(t(Samples_n)), Avg_ftn, by=0) %>%
  filter(Bg_bin == 0) %>% #picking only the features
  dplyr::select(-c(Avg_blank,Avg_samples,Ratio_blank_Sample,Bg_bin)) #removing the last 4 columns
row.names(blk_rem_n) <- blk_rem_n$Row.names #extra step bc columns-to-rownames didn't work
blk_rem_n <- as.data.frame(t(blk_rem_n %>% dplyr::select(-c(Row.names))))
dim(blk_rem_n) #now the samples are retained
names(blk_rem_n) #and so are the metabolite names :)

```


```{r maaslin2 baseline input}
#repeat on the BLNK REMOVED DATA - skip imputation, this was a mess - baseline only
mdp_clean <- mdp %>% filter(mdp$Metabo_SampleID %in% row.names(blk_rem_p)
                            & Blood=="B1" & !is.na(Age)) #baseline real samples only!
blk_rem_p <- blk_rem_p[which(row.names(blk_rem_p)%in%row.names(mdp_clean)),]
identical(row.names(mdp_clean), row.names(blk_rem_p))
blk_rem_pt <- as.data.frame(t(blk_rem_p))
print("identical(row.names(mdp_clean), names(blk_rem_pt))") #for log file
identical(row.names(mdp_clean), names(blk_rem_pt))
row.names(blk_rem_pt) #and the metabolite ids are still retained.

mdn_clean <- mdn %>% filter(mdn$Metabo_SampleID %in% row.names(blk_rem_n)
                            &Blood=="B1" & !is.na(Age) ) #baseline samples only!
blk_rem_n <- blk_rem_n[which(row.names(blk_rem_n)%in%row.names(mdn_clean)),]
identical(row.names(mdn_clean), row.names(blk_rem_n))
blk_rem_nt <- as.data.frame(t(blk_rem_n))
print("identical(row.names(mdn_clean), names(blk_rem_nt))")
identical(row.names(mdn_clean), names(blk_rem_nt))
row.names(blk_rem_nt) #and the metabolite ids are still retained.

#write these exact files out for running on the server.
write.csv(mdp_clean, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_input/mdp_clean.csv")
write.csv(mdn_clean, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_input/mdn_clean.csv")
write.csv(blk_rem_pt, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_input/blk_rem_pt.csv")
write.csv(blk_rem_nt, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_input/blk_rem_nt.csv")

```


```{r maaslin2 no covariates}
#repeat on the regular data manually curated by Dai Long
mdp_clean <- mdp %>% filter(mdp$Participant_ID %in% meta_seq$Participant_ID #make sure it's real samples only
                            & Blood=="B1") #baseline samples only!
ftp_t <- ftp_t[which(row.names(ftp_t)%in%row.names(mdp_clean)),]
identical(row.names(mdp_clean), row.names(ftp_t))
ftp <- as.data.frame(t(ftp_t))
identical(row.names(mdp_clean), names(ftp))
row.names(ftp) #and the metabolite ids are still retained.

mdp_clean$fever7d_aboveavg <- factor(mdp_clean$fever7d_aboveavg, levels=c("lo", "hi"))

#write out the cleaned baseline metabolomics data as input for other analysis, e.g. omics integration
#before writing it out, transform metabolomics IDs to participant IDs
identical(mdp_clean$Metabo_SampleID, row.names(ftp_t))
row.names(ftp_t) <- mdp_clean$Participant_ID
write.csv(ftp_t, "/ebio/abt3_projects2/uHEAT2/data/metabolomics/Metabolomics_Pos_B1_Clean_Nov2024.csv")

#fever7d (pos data) - no covariates
fit_data = Maaslin2(
  input_data = ftp, 
  input_metadata = mdp_clean, 
  output = "maaslin2_output", 
  fixed_effects = c("fever7d_aboveavg"),
  normalization="tss", #returned to default recommendation
  transform="log", #& log transform
  #min_abundance=0.0001, #no reasonable min abund exists
  min_prevalence=0.25, #must be present in 25% of samples
  max_significance=0.05, 
  plot_scatter=FALSE
)
pos_fever <- fit_data$results
pos_fever %>% filter(qval<0.1) #centered scaled data also no association, this seems fair.
pos_fever %>% filter(pval<0.05) 
write.csv(pos_fever, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns/maaslin2_output/MetaboPos_Maaslin2_Fever7d_CleanedInput.csv")

##NEG
mdn_clean <- mdn %>% filter(mdn$Participant_ID %in% meta_seq$Participant_ID #make sure it's real samples only
                            &Blood=="B1") #baseline samples only!
ftn_t <- ftn_t[which(row.names(ftn_t)%in%row.names(mdn_clean)),]
identical(row.names(mdn_clean), row.names(ftn_t))
ftn <- as.data.frame(t(ftn_t))
identical(row.names(mdn_clean), names(ftn))
row.names(ftn) #and the metabolite ids are still retained.

mdn_clean$fever7d_aboveavg <- factor(mdn_clean$fever7d_aboveavg, levels=c("lo", "hi"))

#write out the cleaned baseline metabolomics data as input for other analysis, e.g. omics integration
identical(mdn_clean$Metabo_SampleID, row.names(ftn_t))
row.names(ftn_t) <- mdn_clean$Participant_ID
write.csv(ftn_t, "/ebio/abt3_projects2/uHEAT2/data/metabolomics/Metabolomics_Neg_B1_Clean_Nov2024.csv") 

#fever7d (neg data) - no covariates
fit_data = Maaslin2(
  input_data = ftn, 
  input_metadata = mdn_clean, 
  output = "maaslin2_output", 
  fixed_effects = c("fever7d_aboveavg"),
  normalization="tss", #returned to default recommendation
  transform="log", #& log transform
  #min_abundance=0.0001, #no reasonable min abund exists
  min_prevalence=0.25, #must be present in 25% of samples
  max_significance=0.05, 
  plot_scatter=FALSE
)
neg_fever <- fit_data$results #no hits
neg_fever %>% filter(qval<0.1) #no hits
neg_fever %>% filter(pval<0.05) #nothing even remotely impressive, actually
write.csv(neg_fever, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns/maaslin2_output/MetaboNeg_Maaslin2_Fever7d_CleanedInput.csv")

```

```{r maaslin2 with covariates sex age}

identical(row.names(mdp_clean), names(ftp))
#fever7d (pos data) - with covariates
fit_data = Maaslin2(
  input_data = ftp, 
  input_metadata = mdp_clean, 
  output = "maaslin2_output", 
  fixed_effects = c("Sex", "Age", "fever7d_aboveavg"),
  reference=c("Sex", "m"), #tell it to use male as ref for Sex
  normalization="tss", #returned to default recommendation
  transform="log", #& log transform
  #min_abundance=0.0001, #no reasonable min abund exists
  min_prevalence=0.25, #must be present in 25% of samples
  max_significance=0.05, 
  plot_scatter=FALSE
)
pos_fever_corr <- fit_data$results
pos_fever_corr %>% filter(qval<0.05)
pos_fever_corr %>% filter(qval<0.05&metadata=="fever7d_aboveavg") 

write.csv(pos_fever_corr, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns/maaslin2_output/MetaboPos_Maaslin2_SexAgeFever7d_CleanedInput.csv")

identical(row.names(mdn_clean), names(ftn))
#fever7d (neg data) - with covariates
fit_data = Maaslin2(
  input_data = ftn, 
  input_metadata = mdn_clean, 
  output = "maaslin2_output", 
  fixed_effects = c("Sex", "Age", "fever7d_aboveavg"),
  reference=c("Sex", "m"), 
  normalization="tss", #returned to default recommendation
  transform="log", #& log transform
  #min_abundance=0.0001, #no reasonable min abund exists
  min_prevalence=0.25, #must be present in 25% of samples
  max_significance=0.05, 
  plot_scatter=FALSE
)
neg_fever_corr <- fit_data$results #no hits
neg_fever_corr %>% filter(qval<0.05) #no hits
neg_fever_corr %>% filter(qval<0.05&metadata=="fever7d_aboveavg") 

write.csv(neg_fever_corr, "/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns/maaslin2_output/MetaboNeg_Maaslin2_SexAgeFever7d_CleanedInput.csv")

```


```{r plot volcano sex age fever}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
#neg_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeFever7d_CleanedInput.csv", row.names=1)
#pos_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeFever7d_CleanedInput.csv", row.names=1)

neg_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeFever7d.csv", row.names=1)
pos_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeFever7d.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="fever7d_aboveavg") 
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="fever7d_aboveavg") 
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("deepskyblue4", "deepskyblue3", "firebrick4", "firebrick2", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)


#pdf(width=5, height=3.5, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/2B.MetaboVolcanoSexAgeFever_ManAnn.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: Fever-hi vs lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 30, segment.size=0.2, show.legend = FALSE)
#    geom_text(size=2)
p
#dev.off()

#double check annotations for significant hits
hits <- vdata %>% filter(qval<0.1) %>% as.data.frame() #no
annotations_pos %>% filter(id %in% hits$id)
annotations_neg %>% filter(id %in% hits$id)
#indeed no more data is available
#export the hits for manual annotation via HMDB
#write.csv(hits, "/ebio/abt3_projects2/uHEAT/code/R_code/Rscripts_server/maaslin2_output/Metabo_Maaslin2_SexAgeFever7d_hitsforannotatio#n2.csv")

```

```{r plot volcano sex age IgG1 fever}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
#neg_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeFever7d_CleanedInput.csv", row.names=1)
#pos_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeFever7d_CleanedInput.csv", row.names=1)

neg_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeIgG1Fever7d.csv", row.names=1)
pos_fever <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeIgG1Fever7d.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="fever7d_aboveavg") 
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="fever7d_aboveavg") 
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("deepskyblue4", "deepskyblue3", "firebrick4", "firebrick2", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)


#pdf(width=5, height=3.5, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/2E.MetaboVolcanoSexAgeIgGFever_ManAnn.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: Fever-hi vs lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 30, segment.size=0.2, show.legend = FALSE)
#    geom_text(size=2)
p
#dev.off()

#double check annotations for significant hits
hits <- vdata %>% filter(qval<0.1) %>% as.data.frame() #no
#annotations_pos %>% filter(id %in% hits$id)
#annotations_neg %>% filter(id %in% hits$id)
#indeed no more data is available
#export the hits for supplementary table
write.csv(hits, "/ebio/abt3_projects2/uHEAT/code/R_code/Rscripts_server/maaslin2_output/Metabo_Maaslin2_SexAgeIgG1Fever7d_hitsforSTable.csv")

```

```{r plot volcano sex age diet fever}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
neg_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeDiet3Fever7d.csv", row.names=1)
pos_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeDiet3Fever7d.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever_corr %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="fever7d_aboveavg") #take a few borderline to display
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever_corr %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="fever7d_aboveavg") #take a few borderline to display
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("deepskyblue4", "deepskyblue3", "firebrick4", "firebrick2", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)

#are there any significant labelled ones? 
vdata %>% filter(qval<0.05) %>% as.data.frame() #no

#pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/CleanPlots/2B.MetaboFeverVolcano.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: Fever Hi vs Lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 30, segment.size=0.2, show.legend = FALSE)
#    geom_text(size=2)
p
#dev.off()


```

```{r plot volcano sex age diet IgG1 fever}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
neg_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeIgGDiet3Fever7d.csv", row.names=1)
pos_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeIgGDiet3Fever7d.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever_corr %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="fever7d_aboveavg") #take a few borderline to display
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever_corr %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="fever7d_aboveavg") #take a few borderline to display
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("deepskyblue4", "deepskyblue3", "firebrick4", "firebrick2", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)

#are there any significant labelled ones? 
vdata %>% filter(qval<0.05) %>% as.data.frame() #no

#pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/2B.MetaboFeverVolcano.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: Fever Hi vs Lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 30, segment.size=0.2, show.legend = FALSE)
#    geom_text(size=2)
p
#dev.off()


```


```{r plot volcano sex age diet}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
neg_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeDiet3.csv", row.names=1)
pos_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeDiet3.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever_corr %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="Diet3") #
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever_corr %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="Diet3") #
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("coral3", "pink", "seagreen4", "seagreen2", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)
#specifically for metabolomics plots, get rid of stupid long names
vdata$delabel_clean <- ifelse(
  nchar(vdata$delabel) <= 50 & 
  str_count(vdata$delabel, "-") <= 3,
  vdata$delabel,
  NA
) 

#pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/plots5/S#.MetaboDiet3Volcano.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel_clean)) + 
  ggtitle("Serum Metabolomics: Diet Veg vs Omni") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 10, segment.size=0.2, show.legend = FALSE)
#    geom_text(size=2)
p
#dev.off()

#export the significant hits for table S#
vdata_export <- vdata %>% filter(qval<0.1)
#write.csv(vdata_export, "/ebio/abt3_projects2/uHEAT2/data/metabolomics/Metabo_SexAgeDiet3_HitsForTableS#.csv")


```

```{r plot volcano sex age IgG}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
neg_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeIgGB2.csv", row.names=1)
pos_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeIgGB2.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever_corr %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="IgG_B2_aboveavg") #
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever_corr %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="IgG_B2_aboveavg") #
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("skyblue3", "skyblue1", "orchid4", "orchid3", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)

#keep all labels here because they are sparse anyways

#pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/plots5/S#.MetaboIgGB2Volcano.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: IgG hi vs lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 10, segment.size=0.2, show.legend = FALSE)
p
#dev.off()

#export a version specifically for the legend only for all volcano plots (larger points)
pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/plots5/S#.MetaboIgGB2Volcano_LEGEND.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: IgG hi vs lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  geom_text_repel(size=2,  max.overlaps = 10, segment.size=0.2, show.legend = FALSE) +
    guides(colour = guide_legend(override.aes = list(size=5)))
p
dev.off()

#export the significant hits for table S#
vdata_export <- vdata %>% filter(qval<0.1)
#write.csv(vdata_export, "/ebio/abt3_projects2/uHEAT2/data/metabolomics/Metabo_SexAgeIgG_HitsForTableS#.csv")




```

```{r plot volcano sex age IgG1 IgG2}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
neg_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_SexAgeIgG1IgG2.csv", row.names=1)
pos_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_SexAgeIgG1IgG2.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever_corr %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="IgG_B2_aboveavg") #
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever_corr %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="IgG_B2_aboveavg") #
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)
head(vdata) #so in fact the coefs are swapped, i.e., value='lo' (instead of hi, which I find more intuitive)
vdata$coef <- vdata$coef*-1

mycolours <- c("skyblue3", "skyblue1", "orchid4", "orchid3", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)

#keep all labels here because they are sparse anyways

#pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/S#.MetaboIgGB2Volcano.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(pval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: IgG hi vs lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 10, segment.size=0.2, show.legend = FALSE)
p
#dev.off()

#export the significant hits for table S#
vdata_export <- vdata %>% filter(qval<0.1)
#write.csv(vdata_export, "/ebio/abt3_projects2/uHEAT2/data/metabolomics/Metabo_SexAgeIgG1IgG2_HitsForTableS#.csv")




```


```{r plot volcano sex age IgG subset}
#instead of the tile plot, a more informative volcano plot.
#take ALL hits, plus annotations

#import the maaslin model results if they have been run somewhere else
neg_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboNeg_Maaslin2_subset_SexAgeIgGB2.csv", row.names=1)
pos_fever_corr <- read.csv("/ebio/abt3_projects2/uHEAT/code/R_code/uHEAT_CleanMarkdowns_Diet3/maaslin2_output/MetaboPos_Maaslin2_subset_SexAgeIgGB2.csv", row.names=1)

#retrieve annotations
pos_fever_an <- pos_fever_corr %>% 
  rename(id=feature) %>%
  full_join(ap) %>% #positive annotations
  filter(metadata=="IgG_B2_aboveavg") #
pos_fever_an #entirely un-annotated
pos_fever_an[is.na(pos_fever_an)] <- c("None")

neg_fever_an <- neg_fever_corr %>% 
  rename(id=feature) %>%
  full_join(an) %>% #negative annotations
  filter(metadata=="IgG_B2_aboveavg") #
neg_fever_an #almost entirely un-annotated
neg_fever_an[is.na(neg_fever_an)] <- c("None")

identical(names(pos_fever_an), names(neg_fever_an))
vdata <- rbind(pos_fever_an, neg_fever_an)

mycolours <- c("skyblue3", "skyblue1", "orchid4", "orchid3", "grey")
names(mycolours) <- c("DOWN q<0.05", "DOWN q<0.1", "UP q<0.05", "UP q<0.1",  "NO")
vdata$diffexpressed <- c("NO")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.1] <- c("UP q<0.1")
vdata$diffexpressed[vdata$coef > 0 & vdata$qval<0.05] <- c("UP q<0.05")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.1] <- c("DOWN q<0.1")
vdata$diffexpressed[vdata$coef < 0 & vdata$qval<0.05] <- c("DOWN q<0.05")

# Create a new column "delabel" to de, that will contain the name of genes differentially expressed 
vdata$delabel <- ifelse(vdata$diffexpressed!="NO"&vdata$annotation!="None",
                          yes=vdata$annotation, no=NA)

#add a manual annotation from Dai Long
man_an <- read.csv("/ebio/abt3_projects2/uHEAT2/data/metabolomics/manual_annotations.csv")
vdata <- vdata %>% full_join(man_an)
vdata$delabel <-  ifelse(vdata$diffexpressed!="NO"&vdata$annotation=="None",
                          yes=vdata$Annotation, no=vdata$delabel)

#keep all labels here because they are sparse anyways

#pdf(width=6, height=5, "/ebio/abt3_projects2/uHEAT/data/plots5/S#.MetaboIgGB2Volcano.pdf")
p = ggplot(data=vdata, aes(x=coef, y=-log10(qval), 
                              col=diffexpressed, 
                              label=delabel)) + 
  ggtitle("Serum Metabolomics: IgG hi vs lo") +
  scale_colour_manual(values = mycolours) + 
  geom_point(size=0.3) + 
  theme_bw() +
  guides(fill = guide_legend(override.aes = aes(label = ""))) + 
  geom_text_repel(size=2,  max.overlaps = 10, segment.size=0.2, show.legend = FALSE) +
  geom_hline(yintercept=-log10(0.05), linetype="dashed") +
  geom_hline(yintercept=-log10(0.1), linetype="dashed")
#    geom_text(size=2)
p
#dev.off()


```
