==> 11.GrowthRatesCOPTR.Rmd <==
---
title: "04.GrowthRates"
author: "Kelsey Huus"
date: "2024-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
##Packages
library(dplyr)
library(ggplot2)
#library(data.table)
library(tidyr)
```

## COPTR

```{r coptr import}
coptr <- read.csv("/ebio/abt3_projects2/uHEAT/data/COPTR/CoPTR/PTR.csv", 
                  row.names=1)
names(coptr) <- gsub(".fq", "", names(coptr))

tax <- read.csv("/ebio/abt3_projects2/uHEAT/data/output_kraken_all/kraken/tax_table.csv")

#per-sample metadata #note: mocks, blanks & samples with low read depth already excluded #so are old & infected samples
#updated: this version also excludes SERIES 2 of all double series.
meta_seq <- read.csv("/ebio/abt3_projects2/uHEAT/data/metadata/meta_seq_clean_series1.24.08.18.csv", 
                     row.names=1)
meta_seq$fever7d_aboveavg <- factor(meta_seq$fever7d_aboveavg, levels=c("lo", "hi"))
meta_seq$Bristol_Stool_Score2 <- ifelse(meta_seq$Bristol_Stool_Score2=="bNA", yes=NA, no=meta_seq$Bristol_Stool_Score2) #correct BSS label error

#per-person metadata, matching
meta <- meta_seq %>% filter(!duplicated(Participant_ID)&!is.na(fever7d_aboveavg)) %>%
  dplyr::select(-c(SampleID, Relative_Day, Time0, Time01))
length(meta$Participant_ID) #final number of participants

```

```{r coptr process}
#start with making an average growth rate across all bacteria counted here
coptr$Species <- row.names(coptr)
coptr_l <- coptr %>% pivot_longer(cols=c(1:1194), names_to="SampleID", values_to="PTR")
#add some taxonomic info
tax$Species <- gsub(" ", "_", tax$Species, fixed=TRUE)
coptr_l <- full_join(coptr_l, tax)
#avg across all bacteria
#were archaea functionally detected? no
coptr_l %>% filter(!is.na(PTR)) %>% #now that it's long format, kick out NAs
   dplyr::count(Domain)
coptr_avg <- coptr_l %>% filter(!is.na(PTR)) %>% #now that it's long format, kick out NAs
  group_by(SampleID) %>%
  summarise(Mean_PTR=mean(PTR))
coptr_avg
#avg by phylum
coptr_avg_phylum <- coptr_l %>% filter(!is.na(PTR)) %>% #now that it's long format, kick out NAs
  group_by(SampleID, Phylum) %>%
  summarise(Mean_PTR_phylum=mean(PTR))
coptr_avg_phylum
#avg by family
coptr_avg_fam <- coptr_l %>% filter(!is.na(PTR)) %>% #now that it's long format, kick out NAs
  group_by(SampleID, Family) %>%
  summarise(Mean_PTR_family=mean(PTR))
coptr_avg_fam
#avg by genus
coptr_avg_genus <- coptr_l %>% filter(!is.na(PTR)) %>% #now that it's long format, kick out NAs
  group_by(SampleID, Genus) %>%
  summarise(Mean_PTR_genus=mean(PTR))
coptr_avg_genus

#combine with metadata and check over time
coptr_meta <- full_join(meta_seq, coptr_avg) %>%
  filter(!is.na(fever7d_aboveavg))

```

```{r coptr visualize}
#visualize overall bacterial growth rates over time, by fever
coptr_meta %>% dplyr::select(SampleID, Mean_PTR, Time01, fever7d_aboveavg) %>% head()

#plot all
#cairo_pdf(width=4, height=3, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/S5F.BacteriaPTR.pdf")
p <- ggplot(coptr_meta, aes(x=Time01, y=Mean_PTR, colour=fever7d_aboveavg)) + 
  #geom_point() + 
  geom_line(aes(group=Participant_ID), alpha=0.3) +
  geom_smooth(method='loess', fill='white', alpha=0.7) + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10))) + 
  xlab("Time (days)") + ylab("Mean PTR") + 
  ggtitle("Bacterial Growth Rate") + labs(colour="Fever") + 
  scale_x_continuous(breaks = (seq(-7, 7, by = 7))) + 
  scale_colour_manual(values=c("grey50", "firebrick3")) +
  annotate('segment', x=0, xend=0, y=0.4, yend=1.5, linetype='dashed') +
  annotate('text', x=3, y=1.3, label=c("Time p=0.72\nFever p=0.04\nTime:Fever p=0.53"), hjust=0, size=2.5)
p
#dev.off()

#test statistically via mixed linear model #the same way as for beta diversity 
m1 <- lme(Mean_PTR ~ seq_depth +  Bristol_Stool_Score2 + Time01 + fever7d_aboveavg + Time01*fever7d_aboveavg, random=~1|Participant_ID, data = coptr_meta, 
            na.action=na.exclude, method="ML")
summary(m1) #NS

#is there a difference by time of day / circadian rhythm?
pdata <- coptr_meta %>% filter(!is.na(Time_of_Day))
kruskal.test(pdata$Mean_PTR~as.factor(pdata$Time_of_Day)) #p=0.11
p <- ggplot(pdata, aes(x=as.factor(Time_of_Day), y=Mean_PTR)) + geom_boxplot(outlier.size=NA, outlier.shape=NA)
p <- p + geom_jitter(width=0.1, height=0)
p <- p + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10)))
#p <- p + xlab("Day Relative to Vaccine") + ylab("Mean PTR") + ggtitle("Bacterial Growth Rate")
p

#visualize the main phyla over time, by fever.
#plus individual by phyla
coptr_avg_phylum %>% ungroup() %>% count(Phylum) #check which phyla there are #Actino, Bacteroi, Firm_A are the big ones
#I guess the others (e.g. Proteo) were not abundant/prevalent enough to be kept
coptr_meta_Actino <- coptr_avg_phylum %>% 
  filter(Phylum=="Actinobacteriota") %>% 
  full_join(meta_seq) %>%
  filter(!is.na(fever7d_aboveavg))
coptr_meta_Bacter <- coptr_avg_phylum %>% 
  filter(Phylum=="Bacteroidota") %>% 
  full_join(meta_seq) %>%
  filter(!is.na(fever7d_aboveavg))
coptr_meta_FirmA <- coptr_avg_phylum %>% 
  filter(Phylum=="Firmicutes_A") %>% 
  full_join(meta_seq) %>%
  filter(!is.na(fever7d_aboveavg))

#plot by phylum or family
#pdf(width=4.5, height=3.5, "/ebio/abt3_projects2/uHEAT/data/plots2/PTR_Actino_long.pdf")
p <- ggplot(coptr_meta_Actino, aes(x=Time01, y=Mean_PTR_phylum, colour=fever7d_aboveavg)) + geom_point() + geom_smooth()
p <- p + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10)))
p <- p + xlab("Day Relative to Vaccine") + ylab("Mean PTR") + ggtitle("Actinobacteriota Growth Rate")
p <- p + scale_x_continuous(breaks=c(seq(-8, 8, by=1)))
p <- p + scale_colour_manual(values=c("grey50", "firebrick3"))
p
#dev.off()

#pdf(width=4.5, height=3.5, "/ebio/abt3_projects2/uHEAT/data/plots2/PTR_Bacteroidota_long.pdf")
p <- ggplot(coptr_meta_Bacter, aes(x=Time01, y=Mean_PTR_phylum, colour=fever_aboveavg)) + geom_point() + geom_smooth()
p <- p + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10)))
p <- p + xlab("Day Relative to Vaccine") + ylab("Mean PTR") + ggtitle("Bacteroidota Growth Rate")
p <- p + scale_x_continuous(breaks=c(seq(-8, 8, by=1)))
p <- p + scale_colour_manual(values=c("grey50", "firebrick3"))
p
#dev.off()

#pdf(width=4.5, height=3.5, "/ebio/abt3_projects2/uHEAT/data/plots3/PTR_FirmA_long_fever7d.pdf")
p <- ggplot(coptr_meta_FirmA, aes(x=Time01, y=Mean_PTR_phylum, colour=fever7d_aboveavg)) + geom_point() + geom_smooth()
p <- p + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
               panel.background = element_blank(), axis.line = element_line(colour = "black"),
               axis.title.y = element_text(margin = margin(r = 10)))
p <- p + xlab("Day Relative to Vaccine") + ylab("Mean PTR") + ggtitle("Firmicutes A Growth Rate")
p <- p + scale_x_continuous(breaks=c(seq(-8, 8, by=1)))
p <- p + scale_colour_manual(values=c("grey50", "firebrick3"))
p
#dev.off()


```
