==> 13.MaierLabGrowthCurves.Rmd <==
---
title: "10.MaierLabGrowthCurves"
author: "Kelsey Huus"
date: "2024-08-30"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(ggplot2)
```

## Maier Lab Growth Curves

```{r import}
#for processing of the raw growth curves to find the growth rate, see the code locally on my computer
#otherwise, summary table: 
growthrates <- read.csv("/ebio/abt3_projects2/uHEAT/data/MaierGrowthData/21-09-01-manual-growth-rates-reps.csv", row.names=1)
tax <- read.csv("/ebio/abt3_projects2/uHEAT/data/MaierGrowthData/tax_table_maier.csv")

```

```{r summarize}
#pivot longer
head(growthrates)
growthrates <- pivot_longer(growthrates, cols=c(2:10), #re-calculate the averages later
                            names_to="Name", values_to="Growth_rate")
growthrates$Temperature <- sapply(strsplit(growthrates$Name, "_"), `[`, 2)
growthrates$Rep <- sapply(strsplit(growthrates$Name, "_"), `[`, 1)

#do not filter by replicate, allow between 1-3 replicates per strain per condition (hence code skipped)
#take replicates 1-2 
#gr <- growthrates %>% filter(Rep %in% c("r1", "r2") & !is.na(Growth_rate))
gr <- growthrates %>% filter(!is.na(Growth_rate))

#exclude the pathogens - take only commensals
#final no. of strains and their identities:
tax_info <- dplyr::count(gr, id) %>% full_join(tax) %>% filter(!is.na(n))
#but it doesn't make sense to include the pathogens!
tax_info$species
pathogens <- c("Salmonella_enterica_typhimurium", "Enterococcus_faecalis", "Escherichia_coli_UT189",
               "Vibrio_cholerae", "Shigella_flexneri", "Shigella_sonnei", "Clostridium_difficile", 
               "Yersinia_enterocolitica_WA-314", "Yersinia_pseudotuberculoses")
tax_no_pathogens <- tax_info %>% filter(!species %in% pathogens)
length(tax_no_pathogens$id) #25 different commensals
#but also exclude NT5078, E. coli commensal, which doesn't have 37ºC data

#export tax info for table s2
#write.csv(tax_no_pathogens, "/ebio/abt3_projects2/uHEAT/data/MaierGrowthData/tax_table_maier_filt.csv")

#take a summary R at each temperature, per strain
gr_summary <- gr %>% 
  filter(id %in% tax_no_pathogens$id & id!="NT5078") %>%
  group_by(id, Temperature) %>%
  summarize(Growth_rate=mean(Growth_rate))
gr_summary


```

```{r plot}
#plot the summary growth rates per bacterium per temperature

pdf(width=3, height=3, "/ebio/abt3_projects2/uHEAT/data/RevisionPlots/S6J_invitroGrowthRates.pdf")
p <- ggplot(gr_summary, aes(x=Temperature, y=Growth_rate)) + 
  geom_boxplot(outlier.shape=NA) + 
  geom_point() + 
  geom_line(aes(group=id), alpha=0.2) + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"),
        axis.title.y = element_text(margin = margin(r = 10))) +
  annotate('segment', x=1, xend=3, y=0.7, yend=0.7) + 
  annotate('text', x=2, y=0.75, label='**',size=5) +
  xlab("Temperature (ºC)") + ylab("Growth Rate") + ggtitle("In vitro growth rates")
p
dev.off()

#test
k <- kruskal.test(gr_summary$Growth_rate~as.factor(gr_summary$Temperature)) #p=0.002
#posthoc test, are 37 and 42 different?
dunn.test::dunn.test(x=gr_summary$Growth_rate, g=as.factor(gr_summary$Temperature), method="bh") #benjamini-hochberg ie fdr
#indeed, 32 is different from 37 abd 42, but 37 and 42 are not different (p=0.15)

#add labels by bacterium or phylum?
gr_summary_tax <- gr_summary %>% full_join(tax_info) %>% filter(!is.na(Growth_rate))
gr_summary_tax$label_rna <- ifelse(gr_summary_tax$species %in% c("Bacteroides_thetaiotaomicron", "Prevotella_copri"),
                                   yes=gr_summary_tax$species, no=NA)
gr_summary_tax$label_rna <- gsub("Bacteroides_thetaiotaomicron", "B.theta", gr_summary_tax$label_rna)
gr_summary_tax$label_rna <- gsub("Prevotella_copri", "P.copri", gr_summary_tax$label_rna)

p <- ggplot(gr_summary_tax, aes(x=Temperature, y=Growth_rate)) + 
  geom_boxplot(outlier.shape=NA) + 
  geom_point(aes(colour=label_rna)) + 
  geom_line(aes(group=id, colour=label_rna), alpha=0.2) + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black"),
        axis.title.y = element_text(margin = margin(r = 10))) +
  annotate('segment', x=1, xend=3, y=0.7, yend=0.7) + 
  annotate('text', x=2, y=0.75, label='*') +
  xlab("Temperature (ºC)") + ylab("Growth Rate") + ggtitle("In vitro growth rates")
p


```

```{r q5}
#calculate the 42 vs 37 growth rates i.e. 'q5', per strain
gr_q5 <- gr_summary %>% pivot_wider(names_from=Temperature, values_from=Growth_rate)
gr_q5

gr_q5$Q5 <- gr_q5$`42` / gr_q5$`37`
gr_q5[order(gr_q5$Q5, decreasing=TRUE),]
gr_q5$id <- factor(gr_q5$id, levels=c(gr_q5[order(gr_q5$Q5, decreasing=TRUE),]$id))

p <- ggplot(gr_q5, aes(x=id, y=Q5)) + geom_point()
p

```
